z: introduce Z_ASSERT_DATAEQUAL()
When the test Z_ASSERT_[L]STREQUAL() fails, we write the string in
hexadecimal when there is at least one non printable character in it.

Problems:

* When one string contains a non-printable character and the other
  doesn't, only one of the two strings are printed in hexadecimal,
  making it hard to see what differs.

* Strings that are almost totally printable are written in hexadecimal
  anyway.

* UTF-8 strings are badly handled and usually printed in hexadecimal.

When testing, we usually know whether we expect the strings to be
printable or not, so this patch introduces Z_ASSERT_DATAEQUAL(), that
prints both strings in hexadecimal.

The behavior of the function is changed so only the invalid UTF-8
characters are printed in hexadecimal preceded by the escaping '\x'.

Note: as 'sb_put_string()' changes behavior, stop using it in
'_z_assert_iopjsonequal()' and simply print the json with sb_add_lstr():
the escaping of '\n' characters made the JSON unreadable anyway.

Change-Id: I0c505d085ab68d3b3920f69260b221a04e9ef067