log: Quote logger names in "leaked logger" message.
Quoting names is mandatory in log messages, because in many cases we do
not understand that the printed names are variable contents, and miss
the real meaning of the message.

In this case, this is particularly true when "logger" is the root
logger. We used to have:

 leaked logger fake, cannot wipe

And we'll now have:

 leaked logger `fake`, cannot wipe ``

Change-Id: Ie79aadaae7fe4380d51e811af07b19f051425b76