iopy: fix deprecated implicit conversion using __int__ with enum.
With Python 3.8+ a warning was raised when implicitly converting an enum
to an int:
Implicit conversion to integers using __int__ is deprecated, and may be
removed in a future version of Python.

The solution is to define the method `__index__` that is the one
actually used when converting a numeric object to an integer.

For Python <= 3.7, __index__() and __int__() must return the same value.

See
https://docs.python.org/3/reference/datamodel.html?highlight=__int__#object.__int__

Change-Id: If317e0da01fa6c299a56df30dcf449c650fa0498