readability: use qv_extend() instead of qv_splice() when appropriate
qv_extend() appends the content of a 'tab + len' structure into a
vector. qv_splice() does a lot more things but is a bit less
straight-forward to read: we often have to consult the definition in
order to be sure of what the call does.

This patch makes the switch from 'qv_splice()' to 'qv_extend()' in all
situations where we use 'qv_splice()' in places where 'qv_extend()'
would do just as well.

Purpose: slightly improve the readability of the concerned code.

Refs: #79937
Change-Id: Ib253c0773c6dd129ae43bb3211b252227895d64b