iopc: rethink representation of arg/res/exn in iopc_fun_t
The way the arguments/result/exception were represented in iopc_fun_t
before this patch was inconvenient as the same three attributes were
duplicated three times.

A consequence of this poor design choice was that the developers had to
create big and hardly readable macros when they had to deal with the
attributes.

This patch introduces the struct 'iopc_fun_struct_t' that is used for
both arguments, result and exception so it can reduce the duplication
and improve the code readability as it replaces big macros by cleaner
function calls.

Refs: #79937
Change-Id: I7970bf0433ec416314aea9648ba584995465dfcd