iopc/lang-c: fix folding in *.iop.c, broken by struct aliases
In order to make generated C files readable, IOPC generates folds for
struct, classes, interfaces, etc...

For struct however, it was done in a strange way: the fold was opened in
'iopc_struct_dump_src()' but closes in 'write_struct()'. It was silly
enough to deserve a fix but the bad thing was that 'write_struct()' was
called for each alias of the struct so the fold could be closed several
times.

Refs: #79937
Change-Id: Id9a380d039aba2ef28f55385c166151c9b8935cd