parseopt: introduce options qvector + one dedicated helper
In some advanced cases, the usual static arrays of options are not
sufficient and the options needs to be written into vectors.

This patch declares the qvector type 'qv_t(popt)' for options 'popt_t'
and add a macro that allows to fill vectors of options by using the same
macros in almost the same way.

The new macro 'OPT_VEC_EXTEND_VA()' so as the helper it uses
'opt_vec_extend()' handle the terminating OPT_END() so the user doesn't
need to care about it (quite like 'sb_t' with '\0' but in a slightly
more permissive way).

Change-Id: I1e1ef94c2412de327b4df3f293d07a5b0386f00e