waf/poetry: support poetry in waf configure and build.
We now use Poetry to handle Python dependencies used on compilation
(Cython) and in tests.
`poetry install` is automatically launched on `waf configure`, and when
waf is not run inside the poetry environment, waf is rerun with poetry
on `waf configure` and `waf build`.

Poetry is not used by default when lib-common is used by other projects
unless ctx.use_poetry is set to True before using lib-common.

For projects that use lib-common, it is possible to set callbacks
`before_poetry_install` and `after_poetry_install` to run custom
functions before and after `poetry install`.

Change-Id: I27af9e59bbfe32842dd582359efcdc9c9c20021f