pxcc: handle forward record declaration and recursive registering.
In C, it is possible to do some forward declarations of structs and
unions.
It is also possible in Cython with a similar syntax.

To support this, PXCC needs to know the order of declaration and
definition to avoid unknown types.
Unfortunately, libclang-c does not permit to distinguish between the
forward declaration and the definition of a struct or union.

The solution is to handle the recursive usage of types when we register
the different types, and for structs and unions, generate a forward
declaration when they are used recursively.

Change-Id: Iedcd61a645f88ccfd18de0713f4872a7ef1811a4
Refs: #65559