container/qvector: introduce 'qv_extend(vec, ptr, len);'
Before this patch, we had a very convenient 'qv_extend()' that allows to
append the content of an array-like structure into a qvector.

This was neat but we had no equivalent that takes a (pointer + length).
This patch renames the existing 'qv_extend()' as 'qv_extend_tab()' and
introduces 'qv_extend()': the (pointer + length) equivalent.

Also:

* Make a pass on existing uses of 'qv_splice()' to replace them with
  'qv_extend()' (or 'qv_remove()' ^^) when appropriate.

* Prefix the local variables in the macros with the macro name in order
  to decrease the chances of name conflicts.

Change-Id: Ie9e2f4ea54ff9a4370fc1d760d05708a73e86a23