iopy: Display the name of the abstract class in error message.
The error message "cannot instantiate an abstract class" was not
displaying the name of the class we try to instantiate, making it quite
hard to debug when it happens.

Also change the error messages so that they do not start by an uppercase
letter, as imposed by our coding rules.

Change-Id: Ie75d110d26ed1f3ea840d5300f844eba520510ae