clang-rewrite-blocks: import and compile RewriteBlocks.
In order to rewrite .blk to .blk.c files, and .blkk to *.blkk.cpp files,
we use a special modified clang that have a command `-rewrite-blocks`:
https://github.com/Intersec/clang

This command is handled by RewriteBlocks.cpp.

In order to have a better open-source support and avoid compiling a
whole clang for this option, we can have a standalone
`clang-rewrite-blocks` that uses clang as a lib.

This patch introduces RewriteBlocks.cpp with an empty main(), and links
clang-rewrite-blocks against lib clang and llvm.

Currently, RewriteBlocks.cpp only supports llvm 9.

Change-Id: I434b878c472e3e264661966395aab3800c23b156
Refs: #80732