cflags: fix compatibility with Clang 11.
On Clang 11, `clang --version` returns the platform as a prefix,
e.g.  Debian clang version 11.0.1-2.
This means that the grep used to get the clang_version failed because
the 3rd argument of the line would `version`.

The solution is to use the only matching argument `-o` for grep with a
finer regex.

Change-Id: Id4cfb0860d6216b58a3b720396aeec6128225084
Refs: #80732