parseopt: fix global-buffer-overflow on copyinit().
When copying the initial opt value in copyinits(), for FLAG and INT, we
just cast the value into an int.
This is obviously wrong since for FLAG and INT, the value can be of
different size and is not necessarily an int.
Moreover, UINT handling was also missing and should be handled like FLAG
or INT.

The solution is to do just like put_int_value(), and use
`opt->int_vsize` to get the correct type of the value.

This bug was not seen before because usually, the value points to a
field of a global structure where all the fields are initialized with 0.
So even if the cast is done with an int of the wrong size, we still get
0.
The issue was found on 2018.3 because `iopc_do_typescript_g` is a struct
with only a bool field of size 1. So accessing this field as an int of
size 4 triggers a `global-buffer-overflow`.

The issue was found with Clang 12 Asan:

==2704851==ERROR: AddressSanitizer: global-buffer-overflow on address 0x00000130ea80 at pc 0x0000006bf504 bp 0x7ffee872c930 sp 0x7ffee872c928
READ of size 4 at 0x00000130ea80 thread T0
    #0 0x6bf503 in copyinits /home/pauss/dev/lib-common2/parseopt.c:234:26
    #1 0x6bec76 in parseopt /home/pauss/dev/lib-common2/parseopt.c:256:5
    #2 0x4fd6dd in main /home/pauss/dev/lib-common2/iopc/iopc.c:235:12
    #3 0x7f4fe9839d09 in __libc_start_main csu/../csu/libc-start.c:308:16
    #4 0x422979 in _start (/home/pauss/dev/lib-common2/iopc/iopc+0x422979)

0x00000130ea81 is located 0 bytes to the right of global variable 'iopc_do_typescript_g' defined in 'iopc/iopc-lang-typescript.c:16:33' (0x130ea80) of size 1
  'iopc_do_typescript_g' is ascii string ''
SUMMARY: AddressSanitizer: global-buffer-overflow /home/pauss/dev/lib-common2/parseopt.c:234:26 in copyinits
Shadow bytes around the buggy address:
  0x000080259d00: 00 f9 f9 f9 f9 f9 f9 f9 01 f9 f9 f9 f9 f9 f9 f9
  0x000080259d10: 00 f9 f9 f9 f9 f9 f9 f9 01 f9 f9 f9 f9 f9 f9 f9
  0x000080259d20: 00 f9 f9 f9 f9 f9 f9 f9 01 f9 f9 f9 f9 f9 f9 f9
  0x000080259d30: 00 f9 f9 f9 f9 f9 f9 f9 01 f9 f9 f9 f9 f9 f9 f9
  0x000080259d40: 00 f9 f9 f9 f9 f9 f9 f9 00 00 00 f9 f9 f9 f9 f9
=>0x000080259d50:[01]f9 f9 f9 f9 f9 f9 f9 00 00 00 f9 f9 f9 f9 f9
  0x000080259d60: 01 f9 f9 f9 f9 f9 f9 f9 00 f9 f9 f9 f9 f9 f9 f9
  0x000080259d70: 01 f9 f9 f9 f9 f9 f9 f9 00 f9 f9 f9 f9 f9 f9 f9
  0x000080259d80: 01 f9 f9 f9 f9 f9 f9 f9 00 f9 f9 f9 f9 f9 f9 f9
  0x000080259d90: 01 f9 f9 f9 f9 f9 f9 f9 00 f9 f9 f9 f9 f9 f9 f9
  0x000080259da0: 01 f9 f9 f9 f9 f9 f9 f9 00 f9 f9 f9 f9 f9 f9 f9
Shadow byte legend (one shadow byte represents 8 application bytes):
  Addressable:           00
  Partially addressable: 01 02 03 04 05 06 07
  Heap left redzone:       fa
  Freed heap region:       fd
  Stack left redzone:      f1
  Stack mid redzone:       f2
  Stack right redzone:     f3
  Stack after return:      f5
  Stack use after scope:   f8
  Global redzone:          f9
  Global init order:       f6
  Poisoned by user:        f7
  Container overflow:      fc
  Array cookie:            ac
  Intra object redzone:    bb
  ASan internal:           fe
  Left alloca redzone:     ca
  Right alloca redzone:    cb
  Shadow gap:              cc
==2704851==ABORTING
Aborted
 IOP iop/tstiop_rpc.iop

Change-Id: I86841dc2996723abafe9c582a51dd0af191d20f7