zchk/parseopt: check '--no-' args and initial copy of correct int size.
This commit adds a non-regression test for 99334d2841229 and
4c4670dbece02.

The only way to reproduce the bugs are to use the "--no-" feature
because it is the only use case for field "init".

It checks that all bytes of int of all sizes are well copied in the
init field of an option, and copied back when using an unset argument
for the option.

Change-Id: I1f34043e55c078bdc1c3d91e2604e7af47f09869