parseopt: fix using get_int_init() for OPTION_FLAG.
When OPTION_FLAG is used, `opt->init_vsize` is set to 0.
However, get_int_init() expects a init_vsize to correspond to a real int
size.
So we should not use get_int_init() in that case, otherwise we would get
a panic "should not happen".

In fact, `opt->init` is not used in case of OPTION_FLAG, so we can skip
setting it in copyinits().

Change-Id: I0eee8ba8f4fcf19d93dee2ac44eddacb90f16e5a