iop: rpc-channel, add async queries handling for local ichannel
Currently when using a local ichannel, queries are always done
synchronously. While this is fine most of the time, this can have
unwanted side effects when a module uses both local and remote
ichannel.
This patch allows to configure a local ichannel to use async queries
handling. In that case all IOP queries on a local ichannel are queued
like in the case of remote channel. The incoming queries and
replies are flushed using a dedicated timer.

Refs: #80682
Change-Id: I7ea3a076f06d677827ea84c7922d91569ba96904