core/obj: rework obj_release()
(cherry picked from commit 0b8ed76edb5c5e1c7a5f68fe85167cc17bcb498b)

The macro obj_release() as it was implemented before this patch had an
important drawback: it didn't clear the object pointer, leading to two
awkward consequences:

1. Using an object pointer after it was destroyed by obj_release()
   didn't necessarily led to an immediate crash:

    obj_release(obj);
    printf("%d", obj->foo); /* Invalid read. Potentially no crash. */

2. If the user wanted to know if the object was going to be destroyed,
   he needed to check the 'refcnt' attribute, breaking the
   encapsulation and led to some awkward code:

    ret = ctx->refcnt > 1 ? ctx : NULL;
    obj_release(ctx);

    return ret;

This patch changes the obj_release() macro, making it a bit more like
obj_delete() at is takes a double pointer. The main difference is that
it still cannot take NULL objects and it doesn't reset the pointer if
the object isn't destroyed so that the caller has a way to know if the
object has been destroyed by the release without having to check the
reference counter.

The two examples above are improved:

1.
    obj_release(&obj);
    printf("%d", obj->foo); /* Immediate crash if the object is
                               destroyed. */

2.
    obj_release(&ctx);
    return ctx;

Change-Id: Ie2fcf743cb3138599737c60a0791b80cce19e1be