lib-common/net/http: fix invalid status used in case of timeout.
When the client http has not received events for a while (noact_delay),
EL_EVENTS_NOACT is sent by the event loop, and we should return
HTTPC_STATUS_TIMEOUT in that case according to d9c36ee140b.

Unfortunately, due to an wrong merge conflict resolution in
ccf359053516 (conflict with 35bb0c743653), setting the status to
HTTPC_STATUS_TIMEOUT was lost in case of EL_EVENTS_NOACT.

The solution is to properly set the status to HTTPC_STATUS_TIMEOUT in
case of EL_EVENTS_NOACT when 100cont is not handled.

Change-Id: I77ab37b88d00d9f6b9beb639dfa9b3fcdaf0b904
Refs: #81023