iopc/lang-c: get rid of macro DUMP_ATTRS()
Improve maintainability of iopc-langc by removing an awful macro.
Also rework the code that was in the macro for better readability.

In order to make this patch possible, I had to introduce the enumeration
"iopc_attr_kind_t" which should have a wider use in the next patches to
come, as we push forward the transition to a more generic approach of
IOP attributes.

Change-Id: If836c9d47fbf6637c4fe193537cd38f53b3bca4e