qps-bitmap: Fix possible crash in delete_nodes.
We observed a few times production databases that crashed in this code.

We assumed that when it happened, the QPS had been corrupted in some
ways, and maybe this is true, but this situation is actually easy to
detect, and thus this is possible to avoid crashing.

What was observed is that just before crashing in qps_pg_unmap, the
"assert ((*dispatch)[p].node != 0)" was triggered on debug builds.
Skipping the call of qps_pg_unmap when the node is 0 is not enough as it
seems that after having seen a 0 the next values of "node" are garbage.

So just break when the page is 0, and have an expect to still detect
this invalid situation.

Refs: #82253
Change-Id: Ie2242ac38cc49f9759c0c0fdcc743ad98928b029