iop: rpc-channel, resolve IP address in ic_connect
Before, the IP address resolution was done once for all before the call
to ic_connect. This is a problem with containers because their addresses
can change especially when there is a crash and it restarts.

Now, there is a resolution of the IP address at each connection attempt.
It should be backward compatible.

Refs: #80869
Change-Id: I8e448a8b624c30ec71b81fc87bfd8a2d778a1a78