iopc/lang-c: rework the concept of 'iopc_attr_kind_t'
Merge 'iopc_attr_kind_t' (previously an enumeration) and
'iopc_attr_kind_meta_t' into a single structure and use global instances
of 'iopc_attr_kind_t' to represent all the characteristics of each
attribute kind (and there are plenty).

Purposes:

1. Allow us to decrease the number of parameters in functions that
   dump attributes (we just pass 'kind').

2. Centralize the information to improve the readability.

3. Ease the code deduplication: there are a lot of dumping functions
   that do almost the same thing. We can decrease their number a lot and
   this 'iopc_attr_kind_t' will help.

For now, the 'iopc_attr_kind_t' is not used as much as it can. It's
going to be improved in subsequent patches.

Also add TODO for some of the next steps of the IOPC/lang-c code
cleaning.

Refs: #79937
Change-Id: I2a827447deb1b56f9380b06702b6e6d03d2abc41