str-l: introduce ASCII case converters that are not "in-place"
The existing helpers 'lstr_ascii_tolower()', 'lstr_ascii_toupper()',
'lstr_ascii_reverse()' modify the strings passed to them but most of the
time we have to do a copy of the string because the ownership of the
string is not clear enough to modify it.

This patch adds functions 't_lstr_ascii_{tolower, toupper, reverse}()'
to adapt to most common cases.

Change-Id: I9a76c5b210124d64ba1a4019af70ad0c2811b3c7