iopc/lang-c: simplify some attribute-dumping functions
A lot of functions took two parameters and used them to build variable
names in the C code they dumped. They also deduced implicit things from
the parameter presence: 'iopc_dump_attrs_ref()' knew when it was dumping
elements from an array because the parameter 'fname' was set.

That was unpractical and insane. This patch let us pass a single prefix
for the variable names prefixes.

If a dumped element is part of an array, tell the function with a
boolean "is_in_array".

Change-Id: I52f8d1db5132e44a7195da02f30bc7718d3ba313
Refs: #79937