lib-common: fix `int` variable usage for returned type of qm_reserve().
We should use `uint32_t` instead of `int` when using qm_reserve().

This is actually not a big issue because `int` is a 32 bits integer, and
checking and removing the most-significant, which is the sign-bit for
`int`, does not pose any issue.

However, doing bit operations on a signed integer is not completely safe
and the returned type of qm_reserve() is a `uint32_t`, so we should use
a `uint32_t` instead of a `int`.

Change-Id: Ia2432df687acfbaa623338bc8de3260c46415386