cython: simplify conversion from python str to lstr_t.
Since 01a2cd2323, we are only supporting Python 3.5+.
This means that we are sure the conversion between a python string to a
lstr_t does not require an allocation for the lstr_t thanks to
PyUnicode_AsUTF8AndSize().

The force_alloc parameter can be also be removed since it was only used
in places where we already know if we have a python str or bytes.

Change-Id: If36354dca8ffe9d0f5b339e7bc8d16a609100596
Refs: #80373