lib-common/iop: fix registering dependencies iop packages.
When registering an IOP package, we didn't actually register the
packages dependencies.

This means that it was possible to register an IOP package A that
depends on another IOP package B, and still not be able to get the IOP
objects of the package B through iop_get_obj().

The solution is to register the dependencies of a package on package
registration.

Change-Id: I58aade05865d24a5edf3da4c0b39fe41337655a5