iopc: rename iopc-pretty.c as iopc-common.c
Some functions in 'iopc-pretty.c' are not relative to pretty printing,
in particular the functions relative to "iopc_path_t", that are used to
generate names and path.

Because we need a place for 'iopc_path_t' functions and don't
necessarily need to have a whole C file for 't_pretty_token()', rename
'iopc-pretty.c' as 'iopc-common.c'.

Change-Id: Ie6a7a1cfd085f221676a849de46eddc8af975876