iopy: fix thread starvation and connection timeout handling.
Problem
=======

When using a synchronous operation that is normally asynchronous in the
event loop, like an IChannel connection or a RPC query, we use a
condition and the internal mutex to wait for the operation in the
current thread.

On top of that, we want to have a timeout for the maximum time waiting
for the operation.

In order to do that, we used pthread_cond_timedwait(3) with an absolute
timeout specified.

It was assumed that once the timeout was reached,
pthread_cond_timedwait(3) immediately acquired the mutex, and thus the
timeout was respected.

Unfortunately, this is not the case as the scheduling between threads
and how mutexes are acquired by threads is not "fair" by default.

In the event loop thread, we use the following code to release, yield
the current CPU, and lock-again the mutex:
    pthread_mutex_unlock(&_G.el_mutex);
    sched_yield();
    pthread_mutex_lock(&_G.el_mutex);

This was supposed to be able to allow other threads to lock the mutex.
But in a multi-processors system, the mutex is very often kept by the
same thread, the event loop thread in our case, and other threads cannot
acquire right away.

This meant that pthread_cond_timedwait(3) could wait much longer than
the specified waiting time, and thus the timeout parameter was not
respected.

Solution
========

We actually encountered this issue of thread starvation with the
internal mutex when trying to lock it simply with pthread_mutex_lock(3).

The solution is to use a condition (_G.el_mutex_after_lock_cond) and an
atomic counter (_G.el_mutex_wait_lock_cnt) to signal the event loop
thread that we want to obtain the internal mutex.

So, to solve our timeout issue, we need to call
iopy_inc_el_mutex_wait_lock_cnt() to signal the event loop thread
that we want to lock the internal mutex when the timeout is reached.

To do that, instead of using pthread_cond_timedwait(3), we now use a
el_timer for the timeout condition with a simple boolean, and we can
simply use pthread_cond_wait().

In iopy_wait_thr_cond_sigint(), we actually missed to call
iopy_inc_el_mutex_wait_lock_cnt(), and we had the same issue of thread
starvation in case of SIGINT.

As a side note, pthread_cond_wait_ext() is now only used by
el_loop_wait_after_lock_cond(), and thus can be inlined in this
function.

Test
====

The issue is triggered when an IChannel is trying to connect to a server
that cannot accept connections or take a long time to do so.

To test it, use a fake TCP server that listens to an address and port,
but never accepts connections.
We test that the connection timeout is respected when using a timeout of
1s.

Since this test takes a minimum of 5s, flag it as slow.

Change-Id: Ideb1aede8d9f186b3116050b2b63bd2b61b095c2
Refs: #83247