iopy: add no activity timeout for client channels.
When connecting to a remote platform, the socket, and the IChannel, is
not necessarily detected as disconnected right away when the socket
becomes invalid in case of server shutdown.

To solve this issue, we have in the IChannel a watch activity timer that
can be activated to disconnect the IChannel after a certain time of
inactivity.

Unfortunately, this was not exposed in IOPy.

Expose it in IOPy when creating the IChannel.

Change-Id: I5986340cf0247280817c389fc932d1fafca49b49
Refs: #83247