qps/hat: make QPS hat tests cover debug printers
The QPS hat debug printers are quite complex but had no test insuring
they work properly.

This patch adds a basic test so this code is roughly covered and so we
can detect behavior changes.

In order to do that, we had to make the function support a "stream"
parameter instead of just using stderr.

Refs: #83283
Change-Id: I43436762a138a220dca3e5d0bd4dc46f2f74e627