qps/hat: fix a bug in qhat_enumerator_go_to()
In some cases, the value pointer set in the tree enumerator could target
the wrong value.

This problem was seen while adding the test also added by this commit.

The reason of the problem is the following:

* The 'go_to' function remembers the position of the previous entry
  (old_pos) so it can fixup the value pointer later during the function
  execution (it spares some pointer indirections).

* When the enumerator enters a leaf of the qhat, the value pointer is
  updated so the pointer fixup isn't expected to work, but this case is
  already handled:

      if (old_node.value != QHAT_PATH_NODE(&en->path).value) {
          old_pos = 0;
      }

* The problem is that we go to a given entry that is on the same leaf as
  the previous one, the "enter_leaf" function resets the "value"
  attribute.
  In that case the check doesn't work and the pointer fixup is going to
  be wrong.

I don't really explain why this issue didn't cause more troubles in
user's applications because it's quite easy to trigger.

This patch fixes the problem by checking if the value pointer of the
enumerator was modified: if it is, then it means that the position of
the value was reset to zero.

Another option would have been to detect when we're opening a leaf we're
already in:

@@ -1197,7 +1197,13 @@ void qhat_fix_stored0(qhat_t *hat)
 static void qhat_tree_enumerator_enter_leaf(qhat_tree_enumerator_t *en,
                                             uint32_t key)
 {
-    en->memory = qhat_node_deref(&en->path);
+    qhat_node_const_memory_t memory = qhat_node_deref(&en->path);
+
+    if (memory.raw == en->memory.raw) {
+        qhat_tree_enumerator_find_entry_from(en, key);
+        return;
+    }
+    en->memory = memory;

The patch applies the same change to 'qhat_tree_enumerator_next()'
because there is no reason to have two ways to test the same thing, as
long as it doesn't cost more, but 'qhat_tree_enumerator_next()' wasn't
broken...

Refs: #83283
Change-Id: I62eafa942c4f26f2aac3e770d3b9517ea2f920c8