qps/hat: strengthen fixing up of QPS hat tree enumerator value pointer
In a previous patch, I modified an obscure check to fix a bug detected
with our newly introduced unit test:

 if (en->value != old_value) {
     /* XXX The value pointer has been modified, probably by
      * 'qhat_tree_enumerator_enter_leaf()'. */
     old_pos = 0;
 }

This check (and other existing checks) exists because of the really
unsafe way the enumerator has to fixup the value pointer: the 'value'
pointer is updated while its previous position in the values array is
remembered externally (variable 'old_pos').

The check above is meant to detect when "something changed under our
foot" but it would feel safer to track the position for the value
internally.

This patch adds a 'value_pos' attribute into QPS hat tree enumerator so
we have a better control of the position of the pointer and the
pointer arithmetic we have to do to fix it up is way safer.

Refs: #83283
Change-Id: I99143d601cdd2ce9509db4cdb5ab9c34d8725486