qps/hat: simplification in 'qhat_tree_enumerator_go_to()'
In 'qhat_tree_enumerator_go_to()', we fallback to
'qhat_tree_enumerator_next()' when the target key is the current key
plus one.

That makes sense but we did the fallback only after an important amount
of specific "go_to" code, which doesn't make so much sense.

This patch applies the fallback earlier in the process in order to make
the code more straightforward.

The responsibility of updating the 'value' attribute or ensuring the
safety of the operation is now transferred to the function
'qhat_tree_enumerator_next()', that's we transmit the 'value' and 'safe'
parameters while we did not before this patch.

Refs: #83283
Change-Id: I857545cc59c9dee5ce01c80113a2777ace373071