qps/hat: fix missing 'generation' update after enumerator sync
When a qhat is modified during the lifetime of a qhat enumerator, we
have to re-sync the enumerator before going on with the enumeration.

We're able to know when the enumerator is out-of-sync because the
"generation" in the qhat path maintained in the enumerator stops
matching "qhat.struct_gen".

Problem: the "generation" in the enumerator path is never updated, so we
stayed out-of-sync forever.

Note: the path in the enumerator is a real qhat_path_t but it is all
handled by the enumerator, there are functions to handle a qhat_path_t
but they are never used for this path, and this is probably the reason
why the generation was forgotten.

Change-Id: Ia0168da34464efc6f426ea42115998d5f51453ee
Refs: #83283