qps/hat: fix handling of parameter 'safe' for tree enumerator
The QPS hat trie enumerator supports when the qhat was modified since
the previous enumerator update but it only seems to support a very
limited and inconsistent amount of modifications.

This patch makes the code less awkward by supporting any change in the
compact¹ containing (or supposed to contain) the key.

¹ a compact is a leaf of the trie where all the entries are stored
  consecutively.

I also had to update 'qhat_tree_enumerator_fixup_value_ptr()' because
entry removals may force us to make the value pointer go backward, so
the 'offset' variable had to be signed.

Refs: #83283
Change-Id: Id17cf207fe8de0d918751cc2efef48787dbda04f