qps/hat: check and remove a spurious FIXME about nullable qhats
In a previous documentation patch, I added a FIXME about a possibly
broken pointer: the 'value' pointer in 'qhat_tree_enumerator_t' should
always be a valid pointer because we often shift it when moving from a
value to another.

My mistake that was the value pointer we changed was the value pointer
in 'qhat_enumerator_t', not the one in 'qhat_tree_enumerator_t'.

I could have just removed the FIXME but it feels better to include the
test I specifically coded for this case.

Change-Id: Id95122c6ad8450674d1e1c18073bce12be253606
Refs: #83283