iopy: fix deadlock on process exit while waiting in a thread.
On process exit, we close and join the event thread loop, and clean-up
the thread conditions and mutexes in the main thread.

Unfortunately, if another thread was waiting for some connection or
query, we never notified the thread that the event loop was closing.

This led to a dead lock since the event loop was no longer running and
we can no longer wake up the waiting thread.

The solution is to wake the waiting threads when we close the event
loop, so they can exit properly.

Change-Id: I768d365934f59f05c8d004e0fe390fe17bf10951