qps/hat: add preparatory non-regression test for nullable qhats
Nullable QPS hats rely on QPS bitmaps to know which key is null.

QPS hats enumerators are supposed to have "safe" mode in order to
support when qhats are modified during the lifetime of the enumerator.

QPS bitmaps enumerators don't have "safe" mode but it seems that they
are considered as always safe. But they're not: they seem to handle
changes in the bitmap structure but not changes in the word that is
currently read.

This patch adds a dedicated test to emphasize the problem until it can
be fixed.

Refs: #83283
Change-Id: I3eb5407967c3dea7fa42251c8d60065e288172d1