qps/hat: simplify QPS hat enumerator
QPS hats are a key-value data structure.

Before this patch, the QPS hat enumerator had a hybrid way of handling
values: we could get the value associated to the current key with an
accessor (XXX_get_value/XXX_get_value_safe) and we could as well ask the
enumerator to maintain a 'value' pointer attribute.

Having two different ways to do the same thing doesn't seems so KISS so
this patch removes the 'value' attribute. It keeps an internal
'value_tab' attribute that is set when the enumerator enters a leaf of
the trie (this is already what we did before with 'value') but we don't
maintain a 'value' anymore.

This patch allows us to remove a lot of code. For instance, we don't
have macros 'qhat_for_each_key()' and related variants anymore: all the
macros iterate on the entries and we just have to use the accessor to
get the associated value.

Also add folding in zchk-hat.blk.

Refs: #83283
Change-Id: If562506dc0b0064c99200dce07480b8ccbe7edb3