poetry: remove dependency to numpy and matplotlib.
These two packages were only used by `bench/mem_stats.py` and were not
required to start the tests.

`bench/mem_stats.py` is an old script that is very rarely used.
numpy and matplotlib are huge packages and since we don't really use
them in lib-common, it is best to remove them from the requirements.

Change-Id: I88a8f9107ac38fa2f53f280b591f661ca81fe340