iop/field-path: add explicit class cast syntax for field path.
When filtering/sorting certain objects that uses classes, we sometimes
know that we are using sub-classes for a certain field.

For example with the following IOPs:
    package pkg;

    class MyClass1 {
        int a;
    };

    class MyClass2 {
        int b;
    };

    struct MyStruct {
        MyClass1 cls;
    };

And the following object:
    {
        "cls": {
            "_class": "pkg.MyClass2",
            "a": 1,
            "b": 2
        }
    }

Before, we knew that `cls.b` was accessible, but depending of the
filtering/sorting function used, it was not possible to access it.

We now have a new explicit cast syntax to force considering the class
field as the specified sub-class.
For example, `cls.<pkg.MyClass2>b`.

This would force the field `b` to be considered with the sub-class
`pkg.MyClass2` instead of the base class `pkg.MyClass1`.

Change-Id: Ic0a32b972337fd397931daf9a09719f30b46b4b5
Refs: #83295