iop/field-path: always use on_value callback on iop_field_path_run().
In the next commit we will allow the usage of wildcards for all usages.
In order to do that, we need to always use the callback on
iop_field_path_run() and iop_get_fieldp().

Warning: In iop_get_fieldp(), if we have multiple values matching the
path, we will use the first matching value.
From a user's point of vue, this is now undefined behaviour.

Change-Id: I95c6cac4d8e2eabcca8bc771fcb557939a62cb63
Refs: #83295