httpc/pool: add ability to resolve address on connection.
Add `resolve_on_connect` in `httpc_pool_t` to resolve the address on
connection instead of having to manually resolve the address.

This patch is similar to 8e33039133193dbbd.

This will be used by http_iop_channel_t to only resolve the address on
connection and not when creating the http iop_channel.

Change-Id: I8a1f200a0cbd26c5e6cefc71db4552f54cc8a9a8
Refs: #84006