http_iop_channel: use httpc_pool resolve_on_connect.
Instead of resolving the address manually on http_iop_channel creation,
do it on connection with httpc_pool resolve_on_connect.

This way, if the address is not valid, for example with a failed DNS
resolution, the http_iop_channel is still valid on creation and we will
just try to resolve the address on connection.

Change-Id: I97cbd11fd61db87931cbe7070d22b11472a1b9ea
Refs: #84006