azure-pipelines: fix using older version of clang on ubuntu 18.04.
Clang 9+ is required to compile and use PXCC.
Unfortunately, we still actually used clang 6 to compile lib-common
instead of clang 9.
This caused some compatibility issues.

Instead of adapting the code to keep the compatibility with older
version of clang, use clang 9 in azures pipelines for ubuntu 18.04.

It also makes the azure pipelines more simpler by removing a special
case for ubuntu 18.04.

Change-Id: I74cb3ffc3d3979128d87016072087157a4e22e66