rpc-channel: fix memory leak of unfinished query contexts
When we call a RPC registered with a pre/post hook, the pre_hook will
allocate a context to the query and the post_hook will release this
context. But when the ic module is released with unfinished RPCs, the
post_hook will not be called and therefore the context of queries to
these RPCs will not be released and so they will leak.

This problem occurs since the commit
8cd726f4db6c33a9f7c88f4d7715c1d095729cfd when we started to use the
libc allocator instead of the mem_fifo_pool.
To fix this problem, just release the remaining contexts on shutdown.

Closes: #84986
Change-Id: Ib363324e167f92c333bdb5a5878ae67daf081cd3