qps/bitmap: fix code duplication in bitmap enumerator
The specialization of QPS bitmap enumerator duplicated an important part
of the enumerator code. Splitting it in two parts:

1. The "generic" part, that handles both nullable and non-nullable
   bitmaps.

2. The "non-nullable" part, that duplicates the whole non-nullable part
   of the generic part and the global logic of the implementation.

We could have removed the whole duplication by keeping only the generic
part but the bench shows that it could have implied a small but real
performance drop.

Instead of doing this, this patch splits the code in two "symmetrical"
parts: the non-nullable and the nullable parts.

Consequences:

1. The non-nullable part of the code is no more duplicated.

2. The logic of the implementation is still duplicated. This could be
   fixed using .in.h headers at the cost of a decreased maintainability
   (but the removed duplication would also imply a maintainability
   improvement so it might worth it anyway).

3. The performance of the non-nullable implementation stays the same.
   Creation of a nullable implementation, faster than the generic one.
   The performance of the generic implementation is improved because the
   "is_nullable" attribute is only checked once.

Refs: #83666
Change-Id: I7cbca4052f9830ff72dc64305fe7abdff00579b8