qps/bitmap: get rid of abusive code duplication for enumerator builder
The QPS bitmap enumerator has a lot of specialized code that allows it
to make it slightly faster at the cost of an important duplication.

For now I'm not planning to remove this duplication for the next/go_to
functions because the bench seems to show that this specialization bring
a small performance gain (1-2% of CPU time).

However it really doesn't make sense to keep the duplicated enumerator
builder because no one uses it and we expect the performance gain to be
even smaller: an enumerator can be built once only.

Note: in order to improve code organization and limit the possibility of
conflicts, the non-specialized builder is moved in place of the removed
one.

Also remove the macro 'qps_bitmap_for_each_nn()'. Never used either.

Refs: #83666
Change-Id: Iba26a846649a24023dc81fe1e6f388d58acdd051