qps-bitmap: fix safety issue with QPS bitmap enumerator
Before this patch, the QPS enumerator was supposed to be always safe.
IOW it was supposed to support modifications in a QPS bitmap currently
being enumerated. That didn't really work and that's why the
issue #83666 was opened: the code only handled modifications in the
bitmap structure but some smaller modifications were unsafe .

This patch gets rid of the previous safety-handling code and makes the
bitmap track all its modifications, not only the ones that change the
structure.

It also adds a "safe" parameter on next and go_to APIs, making the
functions faster in cases where we don't expect the bitmap to change.

Contrarily to the older safety code, we only have a single way to
"refresh" our position in the bitmap. It looks like the QPS bitmap
creators had something more efficient in mind but they left absolutely
no documentation about it.

Note: the code specialization between nullable and non-nullable part is
really annoying, making some parts of this patch twice as big as they
should be. I decided to keep the specialization for now but I'll get rid
of it as soon as I can.

====================================
Bench report (ztst-qps-bitmap-bench)
====================================
Tested on my dev laptop. All tests were run several time to eliminate
possible anomalies. All tests were run with -r5 (five scan repetitions).

We're comparing the benchmark result between the patch implementing the
benchmark (old) and this patch (new). The benchmark tool was build with
P=release.

╔══════════╦════════════════╦═══════════════╗
║ bitmap   ║ enumerator     ║   scan time   ║
╟──────────╫─────────┬──────╫───────┬───────╢
║ nullable ║ generic │ safe ║ old   │ new   ║
╠══════════╬═════════╪══════╬═══════╪═══════╣
║ no       ║ yes     │ yes  ║ 3.30s │ 3.15s ║
║          ║ no      │ yes  ║ 3.18s │ 3.15s ║
║          ║ yes     │ no   ║ N/A¹  │ 2.75s ║
╟──────────╫─────────┼──────╫───────┼───────╢
║ yes      ║ yes     │ yes  ║ 4.99s │ 4.62s ║
║          ║ no      │ yes  ║ N/A²  │ 4.60s ║
║          ║ yes     │ no   ║ N/A¹  │ 4.36s ║
╚══════════╩═════════╧══════╩═══════╧═══════╝
¹ There was no unsafe enumerator implementation before this patch.
² There was no nullable enumerator specialization when the benchmark
  was implemented.

In the benchmark context, the new implementation is always faster, and
the addition of the unsafe mode even bring a significant performance
gain. Of course, the context of the bench can be considered as very
different from a real-life use case but these figures are encouraging
and helps us to consider a performance regression as unlikely.

Closes: #83666
Change-Id: I3092396e1cabdc369a9b9aacd7cabbba248077d1