yamlfmt: Add "--delete" option.
When set, and if the input file is in json, it is automatically deleted
after the conversion. This is really useful when repacking a file with
multiple subfiles, to clean the old files that were converted.

This is implemented only when input files are in json format for now
because:
* extracting the list of subfiles from a yaml context does not see to be
  possible without a change in the API
* this is the main use case (conversion of configuration in old json
  format to new yaml)

Change-Id: I1bb376a32336f63396847d01707491aa3bb48334