version.sh: Remove leading zeroes from unsigned versions.
This script parses a git tag, extract from it major/minor and patchlevel
revisions, and dumps them in a C file as unsigned integers.

As a convention, our tags have the number of the month of the release,
on two digits, for the minor revision. So for example, it can be "08" if
the release is made in August.

The problem is that "08" (and "09") is interpreted by the C compiler as
an invalid octal number, leading to subsequent compilation errors when
the generated file is compiled.

Fix that by removing the leading zeroes when parsing the tag.

Change-Id: Ic4245b6fb0d64ec8ed30e6d3e149a5a346f461a8