sctp-tools: support random local port for outgoing connections
The function 'addr_resolve()' is forcing the port to be greater that 1.

When connecting to an SCTP server, we want to be able to not set the
local port and let the system select an unused port.

This patch adds a new function in 'addr.h' that allows the use of the
value 0 to let the system decide which port will be used locally.

This patch applies on the SCTP wrapper the change from commit
b519e11d5bf7b684f.

Change-Id: I042f64093a24448c123fcf7a549b71702bbe0fe1